requirements.txt
pytest
selenium
psycopg2-binary
allure-pytest
python-dotenv
flask
requests